Maintain a local message queue and do Kafka commits instantly
The consumer needs to run commit() on the Kafka queue only when all
messages are consumed. This patch ensures this is done, and preserving
error recovery functionality, by storing received messages on ZooKeeper
before processing them.

If ZooKeeper is not configured, messages will be stored in a local
array. However, error recovery functionality will be lost.