Clean up initialization procedure
The `__main__` module has been refactored into a class, and the
initialization code has been split up into functions. Configuration of
the new `Main` class is done using the `ConfigurableObject`, the same as
for `BaseAdapter`, `BaseExecutor` and `BaseListener`, resulting in a
uniform configuration style.

Furthermore, the splitting of initialization code into functions allows
for the possibility of more test code.