CWFAdapter: fix wrong ProductInstance membership on new resources
If CWFAdapter runs for the first time against a specific reference time,
the Data instances it creates will belong to a wrong ProductInstance,
due to a filtering error.

This commit fixes that problem, and also upgrades productstatus-client
to version 6.1.2, which will throw an exception if this sort of
filtering occurs.