Add NcMLAggregationAdapter.
NcMLAggregationAdapter creates an aggregation NcML file, containing a
set of NetCDF files belonging to the same ProductInstance. The file is
created on the same ServiceBackend as the source files.

This adapter is triggered when ``ncml_aggregation_input_count`` matches
the number of NetCDF files belonging to the current ProductInstance,
Format and ServiceBackend.